Fix signed/unsigned mismatch warnings, part 6.
bias is used only when computing `int32_t e2`, so bias
should also be int32_t in order to avoid compiler warnings.

Also, cast ieeeExponent to int32_t before
computing e2 (which can easily be negative).

+++ ryu/d2s.c
@@ -236,7 +236,7 @@ typedef struct floating_decimal_64 {
 } floating_decimal_64;
 
 static inline floating_decimal_64 d2d(const uint64_t ieeeMantissa, const uint32_t ieeeExponent) {
-  const uint32_t bias = (1u << (DOUBLE_EXPONENT_BITS - 1)) - 1;
+  const int32_t bias = (1u << (DOUBLE_EXPONENT_BITS - 1)) - 1;
 
   int32_t e2;
   uint64_t m2;
@@ -245,7 +245,7 @@ static inline floating_decimal_64 d2d(const uint64_t ieeeMantissa, const uint32_
     e2 = 1 - bias - DOUBLE_MANTISSA_BITS - 2;
     m2 = ieeeMantissa;
   } else {
-    e2 = ieeeExponent - bias - DOUBLE_MANTISSA_BITS - 2;
+    e2 = (int32_t) ieeeExponent - bias - DOUBLE_MANTISSA_BITS - 2;
     m2 = (1ull << DOUBLE_MANTISSA_BITS) | ieeeMantissa;
   }
   const bool even = (m2 & 1) == 0;
